requirements.txt
fastapi
uvicorn[standard]
streamdeck
Pillow
mfrc522
RPi.GPIO
